update Eigen to 3.4; improve performance; add tbb for parallelization

+++ apic2d/math_defs.h
@@ -22,7 +22,7 @@
 #include <Eigen/SparseQR>
 #include <Eigen/StdVector>
 
-typedef double scalar;
+typedef float scalar;
 typedef unsigned long long uint64;
 
 struct int_scalar {
